The FastScanner class has been moved to a common package
