+++ src/main/java/ru/mail/polis/ads/FastScanner.java
@@ -1,4 +1,4 @@
-package ru.mail.polis.ads.part2.marashov.alexander;
+package ru.mail.polis.ads;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -6,15 +6,15 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.StringTokenizer;
 
-class FastScanner {
+public class FastScanner {
     private final BufferedReader reader;
     private StringTokenizer tokenizer;
 
-    FastScanner(final InputStream in) {
+    public FastScanner(final InputStream in) {
         reader = new BufferedReader(new InputStreamReader(in));
     }
 
-    String next() {
+    public String next() {
         while (tokenizer == null || !tokenizer.hasMoreTokens()) {
             try {
                 tokenizer = new StringTokenizer(reader.readLine());
@@ -25,7 +25,7 @@ class FastScanner {
         return tokenizer.nextToken();
     }
 
-    int nextInt() {
+    public int nextInt() {
         return Integer.parseInt(next());
     }
 }
